extract generation methods
